+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update logo ([#72](https://github.com/josefpihrt/orang/pull/72)).
+
 ## [0.5.0] - 2023-09-26
 
 ### Added