+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add globbing functionality ([#74](https://github.com/josefpihrt/orang/pull/74)).
-  - Add options `--include`, `--exclude` 
+  - Add options `--include`, `--exclude`.
+  - Make option `--include-directory` obsolete.
 
 ### Changed