+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0-beta] - 2023-10-19
+
 ### Added
 
 - Add globbing functionality ([#74](https://github.com/josefpihrt/orang/pull/74)).